Ammending my code

+++ two.py
@@ -52,9 +52,9 @@ print("NoneType:", none_var, type(none_var))
 # finally, for, from, global, if, import, in, is, lambda, nonlocal, not, or, pass, raise, return, try, 
 # while, with, yield
 
-a = input("Enter a number: ")
-b = input("Enter another number: ")
-sum = int(a) + int(b)
-print("Sum:", sum)
+a = float(input("Enter a number: "))
+b = float(input("Enter another number: "))
+sum = float(a) + float(b)
+print("Sum:", float(sum))
